- removed old palettes
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@3509 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ modules/file/spmlab.c
@@ -121,7 +121,7 @@ spmlab_load(const gchar *filename)
 {
     GObject *object = NULL;
     guchar *buffer = NULL;
-    guint size = 0;
+    gsize size = 0;
     GError *err = NULL;
     GwyDataField *dfield = NULL;
